fix: parse style colour from either rgb or hex rightly
