--- packages/strapi-plugin-rich-text/admin/src/lib/customElements.ts
@@ -1,8 +1,0 @@
-export function registerCustomElement(
-  elementName: string,
-  constructor: CustomElementConstructor
-): void {
-  if (customElements.get(elementName) === undefined) {
-    customElements.define(elementName, constructor);
-  }
-}
