+++ packages/strapi-plugin-rich-text/admin/src/lib/math.ts
@@ -1,0 +1,7 @@
+export const round = (
+  number: number,
+  digits = 0,
+  base = Math.pow(10, digits)
+): number => {
+  return Math.round(base * number) / base;
+};
